feat: add conda
This was removed in
https://github.com/uktrade/data-workspace/commit/114d9259377f30e4221109679d7523492a1264b6
since it wasn't being used. However, we've now had a request for it (for
multiple users). It's fairly standard in the Data Science world, so let's have
it back.

It should still play well with pip-based installs, so nothing _should_ break

Python 2 I think is a remnant of some older build of theia or node, so should
be safe to remove