build(deps): bump various Theia deps
Will be wanting to bump "multer" later to avoid a vulnerability
https://github.com/expressjs/multer/pull/1097, but bumping everything else now
to hopefully be able to do that one speedily.

This pins to a slightly older Debian image, since I think something's up with
our mirror that causes errors like the below if using a more recent one.

zlib1g-dev : Depends: zlib1g (= 1:1.2.11.dfsg-2) but 1:1.2.11.dfsg-2+deb11u1 is to be installed