build(deps): cut down included Python dependencies
People are using specific environments, e.g. conda environment, for each
project, so most of these dependencies don't end up being used. On the models
we now have a pattern, albeit a bit unscalable, where models can be downloaded
from the mirror. Also suspect that the included models are are most used by a
few people. If they are being used, for both spacy and nltk we also have them
in the mirror that have been used successfully. (There is a chance we don't
have the _exact_ same ones there, but I think it's ok to risk it)

Since people are hitting space limits, deeming it more important to allow them
to work than having an un-awkward way to use models for a few users to get at
their models.

From initial tests, this saves ~2.7GB