feat: disable dashboard from pgAdmin
We see a lot of queries from pgAdmin's dashboard when looking at load on the
database. It's hard to tell if they actually cause a problem, but turning them
off by default should at least clear up a lot of noise.

This includes an increase of the version of pgAdmin just by the nature of
rebuilding with the latest version in PyPI. It's not pinned (it wasn't pinned
before), but I suspect it increases the version from 7.8 to 8.3.