fix: force RProfile reload when rstudio resumes suspended session (#123)
enforcing rstudio session to load in the correct aws credentials when rstudio resumes suspended sessions (sessions are suspended when users trigger shut down on dw tools, or when the user reaches the non-rstudio 2-hour time out, or if a session unexpectedly crashes)

rstudio launched from dw tools does not resume suspended sessions. but rstudio resumes previously suspended sessions when users quit a rstudio session, remains in rstudio, and start a new session. the suspended sessions are on a FILO stack (~/.local/share/rstudio/sessions/active) which pops out when it is resumed. when this stack is empty, rstudio starts a new session. this suspended session was deployed on a now-closed container, which had its own aws access token. this access token is now expired, but rstudio will load it into the r console, which results in http 403 error if users make a call to s3 or aws aurora.

the fix adds an init session hook to rprofile which calls the readRenviron function to load in new env vars which contain the AWS container credential that users need to successfully make request to s3 or aws aurora. init session is an event for initing both a new session and a suspended session. then turning on mandatory rprofile reloads whenever a suspended sessions is resumed, we get the desired effect.

additional: base r update from 4.3.3 to 4.4.0 (no major user-facing changes)