feat: add direnv to the install packages list (and tighten up base image version)
It has been requested by a user to make it easier to manage environment
variables between different projects.

Also removing the specification of LD_LIBRARY_PATH from the JupyterLab
Dockerfile to fix the error:

> LD_LIBRARY_PATH contains the traditional /lib directory, but not the multiarch
> directory /lib/x86_64-linux-gnu. It is not safe to upgrade the C library in
> this situation; please remove the /lib/directory from LD_LIBRARY_PATH and try
> again.

Also taking the opportunity to pin the base image version a bit more to avoid
unexpected changes (or even just wondering if a problem is caused by the image
not being pinned).