Create single Dockerfile for all tools
This moves a lot of the semi-duplicated tools from some of the folders into one
Dockerfile, leveraging multi-stage builds to reduce duplication but still have
ways of having code specific to each tool.

This should make each tool more consistent both with other tools and with
visualisations. There are probably other steps to take after this, this is
"sort of" a refactor, but there are some things that could impact users. For
example, the username for all tools is now "dw-user", when before it was
different for the different tools.

Also taking the opportunity for R to fetch from a major.minor specific location
for compiled R packages, which it should do anyway because only patch versions
are guaranteed to have a stable ABI

Co-authored-by: agalamatis <argyris.galamatis@digital.trade.gov.uk>
Co-authored-by: Sally Mohamed <sally.mohamed@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>