feat: get conda and Python notebooks working in VS Code
Similar to other python-theia, we activate conda when launching a bash login
shell (which needs VS Code specific configuration to start).

Python notebooks are enabled by installing the relevant extensions

Getting Python notebooks working requires a modification to the VS Code /
code-server codebase, to rewrite domains that request resources under
subdomains of vscode-cdn.net to an invalid domain. This rewrite allows us to
open up the CSP on data workspace to this invalid domain, without actually
increasing the risk of data leakage. The reason why this actually helps is
because VS Code has service workers that intercept such requests and responds
with the correct content served internally, without communicating with the
internet.