refactor: s3sync and metrics installed via venv
In order to support the move to bookworm, we install s3sync and the
dependencies for metrics via venv.

Technically we could extend from the python base image as for other
applications, but that would increase the size of the container considerably
because it has a lot of default Python packages installed in it, and so opting
to keep it light.