refactor: really install s3sync in venv
In order to move to Debian bookworm, which is stricter in terms of not
installing Python packages in the root env that are not Debian-y, we need to
install package always via a virtual environment.

In a previous change I thought we did this for s3sync, but I placed the
"activate" in the wrong RUN part of the Dockerfile.