feat: activate conda, increase version, and fix install issues for Python images
This activates conda more consistently for all Python images. Specifically it:

- Sets SHELL so that RUN statements in the Dockefile don't need to each
  activate the conda base environment.

- Sets ENTRYPOINT to use conda run so that the CMD always runs in a conda
  environment.

This is all for moving to bookworm where to install packages via pip, a virtual
environment (like conda) _has_ to be present in general.

The above changes should mean that visualisations don't need to be changed -
they can use python/pip as usual, and they will transparently just run in a
conda environment.

This also does a few things along the way:

- Bumps the version of conda (slightly by the by, but felt should address any
  issues on a more recent conda than an older one that we will have to
  eventually move away from)
- Installs "libmamba", conda's newer/faster solver (not doing this causes some
  odd "module 'libmambapy' has no attribute 'QueryFormat'" errors when
  installing packages, and I suspect slower installs from tools as well)
- Sets the "channel_alias" which better prevents conda from trying to talk to
  the internet in tools (which prevents name resolution failures) Splits the
  conda and Debian package layers (to leverage caching a bit better when
  developing)

These could probably have been released separately, but opting to
not release too many times, so we don't have to fix "transient" issues that
would be fixed by later changes.