feat: hide role and tablespace nodes in pgAdmin
The users don't need to see role and tablespaces in pgAdmin to do data
analysis, so this makes the pgAdmin interface just that tiny bit clearer. And
in fact, since we have many thousands of roles, there are even performance
implications of trying to fetch all the roles, so we should hopefully save some
user/db time from not showing them.